rename: #23 Rename subject service package
Subject service 패키지 이름을 수정하였습니다.
오타가 있었네요..
